+++ src/main/java/com/metalancer/backend/orders/domain/PaymentResponse.java
@@ -13,8 +13,8 @@ public class PaymentResponse {
 
     private final Long ordererId;
     private final String orderNo;
-    private final Integer totalPrice;
-    private final Integer totalPayment;
+    private final Double totalPrice;
+    private final Double totalPayment;
     private final Integer totalPoint;
     private final OrderStatus orderStatus;
     private final String ordererNm;
@@ -24,8 +24,8 @@ public class PaymentResponse {
     private List<OrderProducts> orderProductList;
 
     @Builder
-    public PaymentResponse(Long ordererId, String orderNo, Integer totalPrice,
-        Integer totalPayment, Integer totalPoint, OrderStatus orderStatus, String ordererNm,
+    public PaymentResponse(Long ordererId, String orderNo, Double totalPrice,
+        Double totalPayment, Integer totalPoint, OrderStatus orderStatus, String ordererNm,
         String ordererPhone, String ordererEmail,
         Date purchasedAt) {
         this.ordererId = ordererId;
