+++ src/main/java/com/metalancer/backend/users/domain/PayedOrder.java
@@ -15,13 +15,13 @@ public class PayedOrder {
     private String title;
     private OrderStatus orderStatus;
     private String payMethod;
-    private int price;
+    private Double price;
     private String purchasedAt;
     private String receiptUrl;
 
     @Builder
     public PayedOrder(Long orderId, String orderNo, String title, OrderStatus orderStatus,
-        String payMethod, int price, LocalDateTime purchasedAt,
+        String payMethod, Double price, LocalDateTime purchasedAt,
         String receiptUrl) {
         this.orderId = orderId;
         this.orderNo = orderNo;
